requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.3.0
pyperclip==1.9.0
pystray==0.19.5
qrcode==8.2
six==1.17.0
Werkzeug==3.1.3
